Drop MFA fallback and host sanitizer; records go through the agent.
OTP 27+ already requires voyager_agent on connect, so the dual path, isolate Task, and live tests were overkill. Truncation stays on the agent (VOY-230).
